Migrate data_movement/transpose device operation to TMP pattern
Migration completed:
- [x] Created operation_attributes_t struct with dim, output_mem_config, pad_value
- [x] Created tensor_args_t struct with input tensor reference
- [x] Defined tensor_return_value_t and spec_return_value_t
- [x] Implemented compute_output_specs
- [x] Implemented create_output_tensors
- [x] Implemented select_program_factory returning variant of 6 factories
- [x] Implemented validate_on_program_cache_miss and validate_on_program_cache_hit
- [x] Registered ttnn::prim::transpose
- [x] Updated transpose.cpp to use ttnn::prim::transpose instead of operation::run
- [x] Created 6 program factories with shared_variables_t, create, override_runtime_arguments
- [x] Removed old Transpose struct and transpose_op.hpp/.cpp

Program factories split into separate files:
- transpose_cn_program_factory.hpp/.cpp
- transpose_hc_program_factory.hpp/.cpp
- transpose_hc_sharded_program_factory.hpp/.cpp
- transpose_wh_program_factory.hpp/.cpp
- transpose_wh_sharded_program_factory.hpp/.cpp
- transpose_wh_sharded_rm_program_factory.hpp/.cpp

Resolves: #32761
