+++ ttnn/cpp/ttnn/operations/data_movement/transpose/device/transpose_device_operation_types.hpp
@@ -1,0 +1,31 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "ttnn/tensor/tensor.hpp"
+
+namespace ttnn::operations::data_movement {
+
+enum class TransposeOpDim { WH, HC, CN, NH, NW, CW };
+
+enum class TransposeOpParallelizationStrategy { MULTI_CORE_WH, MULTI_CORE_HC, MULTI_CORE_CN };
+
+namespace transpose {
+
+struct operation_attributes_t {
+    TransposeOpDim dim;
+    tt::tt_metal::MemoryConfig output_mem_config;
+    std::optional<float> pad_value;
+};
+
+struct tensor_args_t {
+    const Tensor& input;
+};
+
+using tensor_return_value_t = Tensor;
+using spec_return_value_t = TensorSpec;
+
+}  // namespace transpose
+
+}  // namespace ttnn::operations::data_movement
