+++ ttnn/cpp/ttnn/operations/data_movement/transpose/device/transpose_program_factory.hpp
@@ -1,17 +1,11 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-//
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace ttnn::operations::data_movement::detail {
+#pragma once
 
-tt::tt_metal::operation::ProgramWithCallbacks transpose_wh_multi_core(const Tensor& a, Tensor& output);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_wh_multi_core_sharded(const Tensor& a, Tensor& output);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_wh_multi_core_sharded_rm(const Tensor& a, Tensor& output);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_hc_multi_core(
-    const Tensor& a, Tensor& output, const std::optional<float>& pad_value);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_hc_multi_core_tiled_interleaved(
-    const Tensor& a, Tensor& output, const std::optional<float>& pad_value);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_hc_multi_core_sharded(const Tensor& a, Tensor& output);
-tt::tt_metal::operation::ProgramWithCallbacks transpose_cn_multi_core(const Tensor& a, Tensor& output);
-
-}  // namespace ttnn::operations::data_movement::detail
+#include "transpose_cn_program_factory.hpp"
+#include "transpose_hc_program_factory.hpp"
+#include "transpose_hc_sharded_program_factory.hpp"
+#include "transpose_wh_program_factory.hpp"
+#include "transpose_wh_sharded_program_factory.hpp"
+#include "transpose_wh_sharded_rm_program_factory.hpp"
