+++ ttnn/cpp/ttnn/operations/data_movement/transpose/transpose.hpp
@@ -1,5 +1,4 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-//
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC.
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
@@ -12,16 +11,13 @@ namespace operations::data_movement {
 struct ExecuteTranspose {
     static ttnn::Tensor invoke(
         const ttnn::Tensor& input_tensor,
-        const int64_t& dim1,
-        const int64_t& dim2,
+        int64_t dim1,
+        int64_t dim2,
         const std::optional<MemoryConfig>& memory_config_arg,
         const std::optional<float>& pad_value = 0.0f);
 
     static ttnn::Tensor invoke(
-        const ttnn::Tensor& input_tensor,
-        const int64_t& dim1,
-        const int64_t& dim2,
-        const std::optional<float>& pad_value = 0.0f);
+        const ttnn::Tensor& input_tensor, int64_t dim1, int64_t dim2, const std::optional<float>& pad_value = 0.0f);
 };
 
 }  // namespace operations::data_movement
